Delete post mutation

+++ src/components/posts/actions.ts
@@ -2,6 +2,7 @@
 import { Post } from "@/generated/prisma/client";
 import { getServerSession } from "@/lib/get-session";
 import { prisma } from "@/lib/prisma";
+import { getPostDataInclude } from "@/lib/types";
 export async function deletePost(id: Post["id"]) {
   try {
     const session = await getServerSession();
@@ -12,10 +13,23 @@ export async function deletePost(id: Post["id"]) {
     if (!id) {
       throw new Error("Invalid inputs");
     }
+    const post = await prisma.post.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!post) {
+      throw new Error("Post not found");
+    }
+
+    if (post.userId !== session.user.id) {
+      throw new Error("Unauthorized");
+    }
     const deletedPost = await prisma.post.delete({
       where: {
         id,
       },
+      include: getPostDataInclude,
     });
     return {
       message: "Post deleted successfully",
